fix to document scroll not responding
What happened: gradient background on <tr> elements with a
background-attachement:fixed css rule prevents the scrolling to work
anymore if there is a css rule overflow:hidden on the root element (html
or body). Funny thing is that applying those css rules _after_ <tr>
elements have been drawn does not trigger the bug. Another workaround is
to set the overflow to auto or scroll. I chose the former solution and
set the scrollbar width to 0, thus having the possibility to scroll,
without having to deal with the ugliness of scrollbars